fix(android): lazy-load Code/Agents so the daemon actually boots (Spike D)
The Code/Agents feature's dependency chain (@earendil-works/pi-coding-agent
-> pi-tui/marked -> typebox) uses \p{...} Unicode-property regex syntax that
TurboLLM Android's embedded (nodejs-mobile) Node build can't even PARSE — a
module-load-time SyntaxError that crashed the whole daemon before any of its
own code ran, even though CodeRunManager's own construction was already
platform-gated. Root cause: esbuild's chunk-splitting shared code across
entries by module-graph reachability rather than feature boundaries, and
`new URL(relative, import.meta.url)` as a dynamic import() specifier gets
statically resolved by esbuild, defeating the lazy load either way.

Fixes, in combination:
- server.ts / cli.ts: construct the dynamic-import specifier via
  pathToFileURL(join(...)) instead of new URL(), which esbuild does not
  recognize as a static asset reference for import().
- tsup.config.ts: split into two genuinely separate build passes (array
  config) so esbuild has no shared-chunk mechanism to leak through at all.
- code/run-buffer.ts: extract RingBuffer/subscribeToBuffer out of
  code-run-manager.ts (which pulls in the whole pi-coding-agent chain) since
  ext/run-manager.ts only ever needed the buffer primitives.
- patches/typebox: durable patch (via patch-package) building 3 ICU-regex
  literals as new RegExp(string) in try/catch instead, so an unsupported
  runtime throws a catchable SyntaxError instead of failing to parse the file.
- research-service.ts: same pattern for the one first-party ICU-regex spot.

Also fixes a second, unrelated crash hit once the above cleared the daemon
to boot further: undici reads a `File` global at its own module-load time
that this embedded Node build doesn't expose (and node:buffer's own File
export is also unavailable here). A plain top-level import in cli.ts is not
early enough to fix this — Node's ESM linker executes CJS dependencies'
top-level code to resolve named exports during the LINK phase, which covers
the whole graph before any module (including the entry's own body) reaches
evaluation. util/android-preload.cjs polyfills File from Blob and is loaded
via a --require preload flag (see the paired TurboLLM-Android commit), which
runs via Node's own CLI bootstrap strictly before it begins loading the
entry module — the only mechanism confirmed to reliably beat that phase.

Verified live on a real Android emulator: the daemon now boots fully and
serves real HTTP responses (GET /api/v1/status, /api/v1/models) on
TurboLLM's own OpenAI-compatible gateway. 3558 tests pass / 4 skip / 0 fail,
typecheck clean.
