+++ turbollm/package.json
@@ -42,8 +42,9 @@
     "typecheck": "tsc --noEmit",
     "test": "tsx --test \"src/**/*.test.ts\" \"web/src/lib/engine-groups.test.ts\" \"web/src/lib/personas.test.ts\" \"web/src/lib/tool-explain.test.ts\" \"web/src/lib/vram.test.ts\"",
     "build:web": "cd web && npm ci && npm run build",
-    "build": "npm run typecheck && tsup && node -e \"const fs=require('fs'),p='dist/cli.js',c=fs.readFileSync(p,'utf8');fs.writeFileSync(p,c.replaceAll('from \\\"sqlite\\\"','from \\\"node:sqlite\\\"'))\" && node -e \"require('fs').cpSync('src/webdist','dist/webdist',{recursive:true,force:true})\"",
-    "prepublishOnly": "npm run build:web && npm run build"
+    "build": "npm run typecheck && tsup && node -e \"const fs=require('fs'),p='dist/cli.js',c=fs.readFileSync(p,'utf8');fs.writeFileSync(p,c.replaceAll('from \\\"sqlite\\\"','from \\\"node:sqlite\\\"'))\" && node -e \"require('fs').cpSync('src/webdist','dist/webdist',{recursive:true,force:true})\" && node -e \"require('fs').cpSync('src/util/android-preload.cjs','dist/android-preload.cjs')\"",
+    "prepublishOnly": "npm run build:web && npm run build",
+    "postinstall": "patch-package"
   },
   "dependencies": {
     "@earendil-works/pi-ai": "0.80.6",
@@ -67,6 +68,7 @@
   "devDependencies": {
     "@types/node": "^22.10.0",
     "@types/sql.js": "^1.4.11",
+    "patch-package": "^8.0.1",
     "tsup": "^8.3.0",
     "tsx": "^4.19.0",
     "typescript": "^5.7.0"
